control: remove unprivileged socket
Instead workout credentials from the connected socket
by getpeereid(2).
Remove the limit of only one connected control client.
Add readgroup, to so we can deny reading of dhcpcd interface
data to everyone on the system.

Always send back a result code to socket callers. This is a breaking ABI change, but allows dhcpcd-online, dhcpcd-gtk etc to display EPERM for example.

+++ src/eloop.c
@@ -236,7 +236,7 @@ eloop_event_count(const struct eloop *eloop)
 static int
 eloop_signal_kqueue(struct eloop *eloop, const int *signals, size_t nsignals)
 {
-	unsigned int cmd = nsignals == 0 ? EV_DELETE : EV_ADD;
+	unsigned short cmd = nsignals == 0 ? EV_DELETE : EV_ADD;
 	struct kevent *ke, *kep;
 	size_t i;
 	int err;
